+++ openeo.d.ts
@@ -1525,6 +1525,7 @@ declare module OpenEO {
          */
         constructor(processes: Array<Process> | Processes, parent?: Builder | null, id?: string);
         /**
+         * List of all process specifications.
          * @type {Array.<Process>}
          */
         processes: Array<Process>;
@@ -1553,6 +1554,13 @@ declare module OpenEO {
          * @type {string}
          */
         public id: string;
+        /**
+         * Adds a process specification to the builder so that it can be used to create a process graph.
+         *
+         * @param {Process} process - Process specification compliant to openEO API
+         * @throws {Error}
+         */
+        addProcessSpec(process: Process): void;
         /**
          * Sets the parent for this Builder.
          *