+++ tests/config.js
@@ -1,5 +1,5 @@
 module.exports = {
 //	TESTBACKEND: 'http://127.0.0.1:8080'
 	TESTBACKEND: 'https://earthengine.openeo.org',
-	STAC_MIGRATE_VERSION: '1.0.0-rc.1'
+	STAC_MIGRATE_VERSION: '1.0.0-rc.2'
 };