+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Custom process specifications can be added `Builder` after first initialization.
+
 ## [1.2.0] - 2021-03-11
 
 ### Added