+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Custom process specifications can be added `Builder` after first initialization.
+- Custom process specifications can be added to `Builder` after first initialization.
+
+### Fixed
+
+- Arrow functions can be used as callbacks in process graph building.
 
 ## [1.2.0] - 2021-03-11